requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langgraph
rank_bm25
jieba
python-dotenv
pypdf